+++ sample/src/androidMain/kotlin/com/github/panpf/sketch/sample/ui/test/TestPage.android.kt
@@ -1,0 +1,3 @@
+package com.github.panpf.sketch.sample.ui.test
+
+actual fun platformTestScreens(): List<TestItem> = emptyList()