+++ sample/src/desktopMain/kotlin/com/github/panpf/sketch/sample/ui/test/TestPage.desktop.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.sample.ui.test
+
+actual fun platformTestScreens(): List<TestItem> = listOf(
+    TestItem("Desktop Temp", DesktopTempTestScreen())
+)